+++ src/commands/index.ts
@@ -1,486 +1,10 @@
-import type {
-  CommandMap,
-  CommandRequest,
-  CommandResult,
-  FollowUpExecutionContext,
-  FollowUpExecutionResult,
-  FollowUpTask,
-} from '@/core';
-import type { AiRuntimeEnv } from '@/skills/ai';
-import { deliverFollowUpEdit, type DeliverFollowUpEnv } from '@/skills/discordInteraction';
-import {
-  EIGHT_BALL_COMMAND_NAME,
-  executeEightBallFollowUp,
-} from '@/commands/8ball';
-import {
-  INSULT_COMMAND_NAME,
-  executeInsultFollowUp,
-} from '@/commands/insult';
-import {
-  PASTIFY_COMMAND_NAME,
-  PASTIFY_MODAL_ID,
-  PASTIFY_MODAL_TEXT_INPUT_ID,
-  executePastifyFollowUp,
-  parsePastifyModalSubmit,
-  type PastifyModalParseResult,
-} from '@/commands/pastify';
-import {
-  ISSUE_BODY_INPUT_ID,
-  ISSUE_COMMAND_NAME,
-  ISSUE_MODAL_ID,
-  ISSUE_TITLE_INPUT_ID,
-  executeIssueFollowUp,
-  parseIssueModalSubmit,
-  type IssueModalParseResult,
-  type IssueRuntimeEnv,
-} from '@/commands/issue';
-import {
-  parseReminderInterval,
-  parseReminderLength,
-  toReminderDelaySeconds,
-} from '@/commands/reminder';
-import {
-  executeReleaseFollowUp,
-  parseReleaseSubcommand,
-  toReleaseSetPayload,
-  RELEASE_COMMAND_NAME,
-  type ReleaseRuntimeEnv,
-} from '@/commands/release';
-
-export { PASTIFY_COMMAND_NAME, PASTIFY_MODAL_ID, PASTIFY_MODAL_TEXT_INPUT_ID } from '@/commands/pastify';
-export { INSULT_COMMAND_NAME } from '@/commands/insult';
-export { EIGHT_BALL_COMMAND_NAME } from '@/commands/8ball';
-export { ISSUE_COMMAND_NAME, ISSUE_MODAL_ID, ISSUE_TITLE_INPUT_ID, ISSUE_BODY_INPUT_ID } from '@/commands/issue';
-export { RELEASE_COMMAND_NAME } from '@/commands/release';
-
-export const WOTD_COMMAND_NAME = 'wotd';
-export const REMINDER_COMMAND_NAME = 'reminder';
-export const SHINY_COMMAND_NAME = 'shiny';
-
-const SHINY_ROLL_MIN = 1;
-const SHINY_ROLL_MAX = 8192;
-
-function rollUniformInteger(minInclusive: number, maxInclusive: number): number {
-  if (!Number.isInteger(minInclusive) || !Number.isInteger(maxInclusive)) {
-    throw new Error('Uniform roll bounds must be integers');
-  }
-
-  const span = maxInclusive - minInclusive + 1;
-  if (!Number.isInteger(span) || span <= 0) {
-    throw new Error('Invalid integer span for uniform roll');
-  }
-
-  const maxUint16 = 0xffff;
-  if (span > maxUint16 + 1) {
-    throw new Error('Uniform roll span exceeds Uint16 sampling domain');
-  }
-
-  const maxAcceptable = Math.floor((maxUint16 + 1) / span) * span - 1;
-  const sample = new Uint16Array(1);
-
-  while (true) {
-    crypto.getRandomValues(sample);
-    const value = sample[0];
-    if (value <= maxAcceptable) {
-      return minInclusive + (value % span);
-    }
-  }
-}
-
-function parseReminderNote(value: string | number | boolean | undefined): string | null {
-  if (typeof value !== 'string') {
-    return null;
-  }
-
-  const note = value.trim();
-  if (!note) {
-    return null;
-  }
-
-  return note;
-}
-
-type ModalComponentRows = Array<{
-  components?: Array<{
-    custom_id?: string;
-    value?: string;
-  }>;
-}>;
-
-function handlePastifyCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command':
-      return {
-        kind: 'show-modal',
-        modalId: PASTIFY_MODAL_ID,
-        title: 'Pastify Idea',
-        inputs: [
-          {
-            inputId: PASTIFY_MODAL_TEXT_INPUT_ID,
-            inputLabel: 'Idea to Pastify',
-            inputPlaceholder: 'Describe the idea to turn into a copypasta',
-            inputMinLength: 1,
-            inputMaxLength: 1000,
-            inputRequired: true,
-            inputStyle: 'paragraph',
-          },
-        ],
-      };
-
-    case 'modal-submit':
-      return {
-        kind: 'enqueue-follow-up',
-        token: request.token,
-        task: {
-          commandName: PASTIFY_COMMAND_NAME,
-          payload: {
-            idea: request.fields[PASTIFY_MODAL_TEXT_INPUT_ID],
-          },
-        },
-        ephemeral: false,
-      };
-
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleInsultCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command':
-      return {
-        kind: 'enqueue-follow-up',
-        token: request.token,
-        task: {
-          commandName: INSULT_COMMAND_NAME,
-          payload: {
-            targetUserId: request.targetId,
-          },
-        },
-        ephemeral: false,
-      };
-
-    case 'modal-submit':
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleEightBallCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command':
-      return {
-        kind: 'enqueue-follow-up',
-        token: request.token,
-        task: {
-          commandName: EIGHT_BALL_COMMAND_NAME,
-          payload: {
-            targetMessageId: request.targetId,
-            targetMessageContent: request.targetMessageContent,
-            targetMessageAuthorId: request.targetMessageAuthorId,
-          },
-        },
-        ephemeral: false,
-      };
-
-    case 'modal-submit':
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleWotdCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command':
-      return {
-        kind: 'ack-and-enqueue-task',
-        content: 'wotd queued',
-        task: {
-          commandName: WOTD_COMMAND_NAME,
-          payload: {},
-        },
-        ephemeral: false,
-      };
-
-    case 'modal-submit':
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleShinyCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command': {
-      const roll = rollUniformInteger(SHINY_ROLL_MIN, SHINY_ROLL_MAX);
-      const content = roll === SHINY_ROLL_MAX
-        ? `You rolled ${SHINY_ROLL_MAX}/${SHINY_ROLL_MAX}. ✨ SHINY ENCOUNTER! ✨`
-        : `You rolled ${roll}/${SHINY_ROLL_MAX}.`;
-
-      return {
-        kind: 'channel-message',
-        content,
-        ephemeral: false,
-      };
-    }
-
-    case 'modal-submit':
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleReminderCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command': {
-      const length = parseReminderLength(request.options.length);
-      const interval = parseReminderInterval(request.options.interval);
-      const note = parseReminderNote(request.options.note);
-      if (!length || !interval || !note) {
-        return {
-          kind: 'channel-message',
-          content: 'Usage: /reminder length:<number> interval:<minutes|hours|days> note:<text>',
-          ephemeral: true,
-        };
-      }
-
-      if (length < 1) {
-        return {
-          kind: 'channel-message',
-          content: 'Length must be at least 1.',
-          ephemeral: true,
-        };
-      }
-
-      if (!request.channelId || !request.userId) {
-        return {
-          kind: 'channel-message',
-          content: 'Could not schedule reminder from this context.',
-          ephemeral: true,
-        };
-      }
-
-      return {
-        kind: 'ack-and-schedule-task',
-        content: `Reminder set for ${length} ${interval}.`,
-        task: {
-          commandName: REMINDER_COMMAND_NAME,
-          payload: {
-            channelId: request.channelId,
-            userId: request.userId,
-            length,
-            interval,
-            note,
-          },
-        },
-        ephemeral: true,
-        delaySeconds: toReminderDelaySeconds(length, interval),
-      };
-    }
-
-    case 'modal-submit':
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleIssueCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command':
-      return {
-        kind: 'show-modal',
-        modalId: ISSUE_MODAL_ID,
-        title: 'Log GitHub Issue',
-        inputs: [
-          {
-            inputId: ISSUE_TITLE_INPUT_ID,
-            inputLabel: 'Issue title',
-            inputPlaceholder: 'Short summary of the bug or feature request',
-            inputMinLength: 1,
-            inputMaxLength: 200,
-            inputRequired: true,
-            inputStyle: 'short',
-          },
-          {
-            inputId: ISSUE_BODY_INPUT_ID,
-            inputLabel: 'Issue details',
-            inputPlaceholder: 'Add any context, repro steps, or desired behavior',
-            inputMinLength: 1,
-            inputMaxLength: 4000,
-            inputRequired: true,
-            inputStyle: 'paragraph',
-          },
-        ],
-      };
-
-    case 'modal-submit':
-      return {
-        kind: 'enqueue-follow-up',
-        token: request.token,
-        task: {
-          commandName: ISSUE_COMMAND_NAME,
-          payload: {
-            title: request.fields[ISSUE_TITLE_INPUT_ID],
-            body: request.fields[ISSUE_BODY_INPUT_ID],
-          },
-        },
-        ephemeral: false,
-      };
-
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-function handleReleaseCommand(request: CommandRequest): CommandResult {
-  switch (request.kind) {
-    case 'command': {
-      const subcommand = parseReleaseSubcommand(request.options.subcommand);
-      if (subcommand === 'list') {
-        return {
-          kind: 'enqueue-follow-up',
-          token: request.token,
-          task: {
-            commandName: RELEASE_COMMAND_NAME,
-            payload: {
-              action: 'list',
-            },
-          },
-          ephemeral: false,
-        };
-      }
-
-      if (subcommand === 'set') {
-        const payload = toReleaseSetPayload(request.options, request.channelId);
-        if (!payload) {
-          return {
-            kind: 'channel-message',
-            content: 'Usage: /release set title:<text> [year:<number>] [quarter:<1-4>] [month:<1-12>] [day:<1-31>]',
-            ephemeral: true,
-          };
-        }
-
-        return {
-          kind: 'enqueue-follow-up',
-          token: request.token,
-          task: {
-            commandName: RELEASE_COMMAND_NAME,
-            payload,
-          },
-          ephemeral: false,
-        };
-      }
-
-      return {
-        kind: 'channel-message',
-        content: 'Usage: /release list or /release set title:<text> [year] [quarter] [month] [day]',
-        ephemeral: true,
-      };
-    }
-
-    case 'modal-submit':
-    case 'component':
-      throw new Error('Unhandled command request');
-
-    default:
-      throw new Error('Unhandled command request');
-  }
-}
-
-export const commands: CommandMap<CommandRequest, CommandResult> = {
-  [PASTIFY_COMMAND_NAME]: handlePastifyCommand,
-  [INSULT_COMMAND_NAME]: handleInsultCommand,
-  [EIGHT_BALL_COMMAND_NAME]: handleEightBallCommand,
-  [ISSUE_COMMAND_NAME]: handleIssueCommand,
-  [WOTD_COMMAND_NAME]: handleWotdCommand,
-  [SHINY_COMMAND_NAME]: handleShinyCommand,
-  [REMINDER_COMMAND_NAME]: handleReminderCommand,
-  [RELEASE_COMMAND_NAME]: handleReleaseCommand,
-};
-
-export function parseCommandModalSubmit(data: {
-  customId: string;
-  components?: ModalComponentRows;
-}): PastifyModalParseResult | IssueModalParseResult {
-  const issueResult = parseIssueModalSubmit(data);
-  if (issueResult.kind !== 'unknown-modal') {
-    return issueResult;
-  }
-
-  return parsePastifyModalSubmit(data);
-}
-
-export async function executeFollowUpTask(
-  task: FollowUpTask,
-  env: AiRuntimeEnv & ReleaseRuntimeEnv,
-  context: FollowUpExecutionContext,
-  issueEnv?: IssueRuntimeEnv,
-): Promise<FollowUpExecutionResult> {
-  if (task.commandName === PASTIFY_COMMAND_NAME) {
-    return executePastifyFollowUp(task, env, context);
-  }
-
-  if (task.commandName === INSULT_COMMAND_NAME) {
-    return executeInsultFollowUp(task, env, context);
-  }
-
-  if (task.commandName === EIGHT_BALL_COMMAND_NAME) {
-    return executeEightBallFollowUp(task, env, context);
-  }
-
-  if (task.commandName === ISSUE_COMMAND_NAME) {
-    if (!issueEnv) {
-      throw new Error('Issue follow-up requires GitHub configuration');
-    }
-    return executeIssueFollowUp(task, issueEnv, context);
-  }
-
-  if (task.commandName === RELEASE_COMMAND_NAME) {
-    return executeReleaseFollowUp(task, env);
-  }
-
-  throw new Error(`Unknown follow-up task command: ${task.commandName}`);
-}
-
-export type FollowUpDeliveryEnv = AiRuntimeEnv & DeliverFollowUpEnv & IssueRuntimeEnv & ReleaseRuntimeEnv;
-
-export async function executeAndDeliverFollowUp(
-  task: FollowUpTask | undefined,
-  context: FollowUpExecutionContext,
-  env: FollowUpDeliveryEnv,
-): Promise<void> {
-  const result: FollowUpExecutionResult = task
-    ? await executeFollowUpTask(
-      task,
-      {
-        AI: env.AI,
-        RELEASES_DB: env.RELEASES_DB,
-        REMINDER_SCHEDULER: env.REMINDER_SCHEDULER,
-      },
-      context,
-      env,
-    )
-    : { content: 'Could not process follow-up payload. Please try again.' };
-
-  await deliverFollowUpEdit(context.token, result, env);
-}
+export * from '@/commands/8ball';
+export * from '@/commands/deploymentStatus';
+export * from '@/commands/insult';
+export * from '@/commands/issue';
+export * from '@/commands/pastify';
+export * from '@/commands/release';
+export * from '@/commands/reminder';
+export * from '@/commands/shiny';
+export * from '@/commands/wotd';
+export * from '@/commands/wordOfDaySchedule';
